# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Driver Connect API",
        "description": "API for managing job categories in the Driver Connect project",
        "contact": {
            "name": "Admin",
            "email": "aleqyanarman@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/get_whatsapp_and_telegram": {
            "get": {
                "tags": [
                    "Social Data"
                ],
                "summary": "Get WhatsApp and Telegram contact information",
                "operationId": "1f1108add1619508ebf6382ee80d3d45",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the operation",
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/job_category": {
            "get": {
                "tags": [
                    "Job Categories"
                ],
                "summary": "Get list of job categories",
                "operationId": "3bee95b93874d816ff938e55a7cc39db",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the operation",
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/get_regions": {
            "get": {
                "tags": [
                    "Regions"
                ],
                "summary": "Get list of regions",
                "operationId": "7d8db2f22bdd64e428dd3825d1ef35e4",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the operation",
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/drive_license_country": {
            "get": {
                "tags": [
                    "Driver License Country"
                ],
                "summary": "Get list of driver license countries",
                "operationId": "0fbe3f0fd03828d05f7dc294e51f1c69",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the operation",
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/car_color": {
            "get": {
                "tags": [
                    "Car"
                ],
                "summary": "Get list of car colors",
                "operationId": "5c97af2a35b8b076da596242167702bf",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the operation",
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/get_car_marks": {
            "get": {
                "tags": [
                    "Car"
                ],
                "summary": "Get list of car marks",
                "operationId": "09edf95999e85b8213f70a2c3304703b",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search string",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the operation",
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/get_car_model": {
            "get": {
                "tags": [
                    "Car"
                ],
                "summary": "Get list of car models by mark ID",
                "operationId": "376f489618874da668a86aa04852f16d",
                "parameters": [
                    {
                        "name": "mark_id",
                        "in": "query",
                        "description": "ID of the car mark",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search string",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the operation",
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the operation",
                                            "type": "boolean"
                                        },
                                        "validation_error": {
                                            "description": "Indicates whether the error is due to validation",
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "description": "Validation error messages",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}